+++ actix-http/src/error.rs
@@ -1021,8 +1021,7 @@ mod tests {
 
     #[test]
     fn test_internal_error() {
-        let err =
-            InternalError::from_response(ParseError::Method, Response::Ok().into());
+        let err = InternalError::from_response(ParseError::Method, Response::ok());
         let resp: Response<Body> = err.error_response();
         assert_eq!(resp.status(), StatusCode::OK);
     }