+++ src/response/mod.rs
@@ -1,0 +1,7 @@
+mod builder;
+mod http_codes;
+#[allow(clippy::module_inception)]
+mod response;
+
+pub use self::builder::HttpResponseBuilder;
+pub use self::response::{CookieIter, HttpResponse};